Split js files in several js files and connect them
